+++ AWS/aws.js
@@ -15,9 +15,6 @@ import {
 import {
   EC2Client,
   GetEbsEncryptionByDefaultCommand,
-  DescribeSnapshotsCommand,
-  DescribeImagesCommand,
-  DescribeFpgaImagesCommand,
   paginateDescribeInstances, paginateDescribeImages, paginateDescribeFpgaImages, paginateDescribeSnapshots
 } from "@aws-sdk/client-ec2";
 import { GlacierClient, paginateListVaults } from "@aws-sdk/client-glacier";
@@ -46,7 +43,7 @@ import {
   S3ControlClient,
   GetPublicAccessBlockCommand
 } from "@aws-sdk/client-s3-control";
-import { SSMClient, ListDocumentsCommand, paginateListDocuments } from "@aws-sdk/client-ssm";
+import { SSMClient, paginateListDocuments } from "@aws-sdk/client-ssm";
 import {
   APIGatewayClient,
   paginateGetRestApis,
@@ -88,11 +85,11 @@ import {
 } from "@aws-sdk/client-compute-optimizer";
 
 import {
-  getAWSAccountId,
   getNextListPageTokenKey,
   getNextListPageTokenRequestKey,
-  getNextPageTokenKeyFromResponse, paginate
-} from "./utils.js";
+  getNextPageTokenKeyFromResponse
+} from "./utils/index";
+import { paginate, getAWSAccountId } from "./service/index";
 import { ListBucketsCommand, S3Client } from "@aws-sdk/client-s3";
 import { EMRClient } from "@aws-sdk/client-emr";
 import { ApiGatewayV2Client, GetApisCommand } from "@aws-sdk/client-apigatewayv2";
@@ -133,7 +130,6 @@ const NEPTUNE_DB_INSTANCE = "AWS::Neptune::DBInstance";
 const NEPTUNE_DB_CLUSTER = "AWS::Neptune::DBCluster";
 const NEPTUNE_DB_CLUSTER_PARAMETER_GROUP =
   "AWS::Neptune::DBClusterParameterGroup";
-const ECS_TASK_DEFINITION = "AWS::ECS::TaskDefinition";
 const WAF_RULE_GROUP = "AWS::WAFv2::RuleGroup";
 const EKS_NODE_GROUP = "AWS::EKS::NodeGroup";
 const COMPUTE_OPTIMIZER_ENROLLMENT_STATUS =
@@ -144,934 +140,46 @@ const S3_BUCKET = "AWS::S3::Bucket";
 const EMR_CLUSTER = "AWS::EMR::Cluster";
 const APIGATEWAYV2_API = "AWS::ApiGatewayV2::API";
 
-const querySSMDocument = async (serviceName, resourceType, region) => {
-  let resources = [];
-  const client = new SSMClient({ region });
-  for await (const document of paginateListDocuments(
-    { client: this.client },
-    {
-      MaxResults: 50,
-      DocumentFilterList: [
-        {
-          key: "Owner",
-          value: "self"
-        }
-      ]
-    }
-  )) {
-    const { DocumentIdentifiers: documentIdentifiers } = document;
-    resources.push(...(documentIdentifiers || []));
-  }
-  const resourceCount = resources.length;
-  updateResourceTypeCounter(serviceName, resourceType, resourceCount);
-  AWS_MAPPING.total += resourceCount;
-};
 
-const queryS3ControlBlockPublicAccess = async (
-  serviceName,
-  resourceType,
-  region
-) => {
-  const client = new S3ControlClient({ region });
-  const command = new GetPublicAccessBlockCommand({
-    AccountId: await getAWSAccountId()
-  });
-  const response = await client.send(command);
-  let total = 0;
-  if (response && response.PublicAccessBlockConfiguration) {
-    total++;
-    updateResourceTypeCounter(serviceName, resourceType, total);
-    AWS_MAPPING.total += total;
-  }
-};
 
-const queryOrganizationsAccount = async (serviceName, resourceType, region) => {
-  let total = 0;
-  const client = new OrganizationsClient({ region });
-  const organization =
-    await client.send(new DescribeOrganizationCommand({}));
-  const resources = [];
-  if (organization.Organization) {
-    resources.push(organization.Organization);
-    total += resources.length;
-  }
-  AWS_MAPPING.total += total;
-};
 
-const queryOrganizations = async (serviceName, resourceType, region) => {
-  const client = new OrganizationsClient({ region });
-  const organization = await client.send(new DescribeOrganizationCommand({}));
-  const resources = [];
-  const roots = [];
-  let total = 0;
-  if (
-    AWS_MAPPING[serviceName] &&
-    AWS_MAPPING[serviceName][resourceType] !== undefined
-  ) {
-    return;
-  }
-  // AWS Organization is only counted as a resource for the master account
-  if (
-    organization &&
-    organization.Organization &&
-    organization.Organization.MasterAccountId === (await getAWSAccountId())
-  ) {
-    resources.push(organization.Organization);
-    updateResourceTypeCounter(serviceName, resourceType, resources.length);
-    total += resources.length;
-    for await (const { Roots: rootPage } of paginateListRoots({ client }, {})) {
-      roots.push(...(rootPage || []));
-    }
-    updateResourceTypeCounter(serviceName, resourceType, resources.length);
-    total += roots.length;
-  }
-  AWS_MAPPING.total += total;
-};
 
-const queryMemoryDBACL = async (serviceName, resourceType, region) => {
-  let total = 0;
-  const client = new MemoryDBClient({ region });
-  const resources = [];
-  let nextToken;
-  do {
-    const command = new DescribeACLsCommand({
-      NextToken: nextToken
-    });
-    const response = await client.send(command);
-    resources.push(...(response.ACLs || []));
-    updateResourceTypeCounter(serviceName, resourceType, resources.length);
-    total += resources.length;
-    nextToken = response.NextToken;
-  } while (nextToken);
-  AWS_MAPPING.total += total;
-};
 
-const queryMemoryDBCluster = async (serviceName, resourceType, region) => {
-  let total = 0;
-  const client = new MemoryDBClient({ region });
-  const resources = [];
-  let nextToken;
-  do {
-    const command = new DescribeClustersCommand({
-      NextToken: nextToken
-    });
-    const response = await client.send(command);
-    resources.push(...(response.Clusters || []));
-    updateResourceTypeCounter(serviceName, resourceType, resources.length);
-    total += resources.length;
-    nextToken = response.NextToken;
-  } while (nextToken);
-  AWS_MAPPING.total += total;
-};
 
-const queryMemoryDBUser = async (serviceName, resourceType, region) => {
-  let total = 0;
-  const client = new MemoryDBClient({ region });
-  const resources = [];
-  let nextToken;
-  do {
-    const command = new DescribeUsersCommand({
-      NextToken: nextToken
-    });
-    const response = await client.send(command);
-    resources.push(...(response.Users || []));
-    updateResourceTypeCounter(serviceName, resourceType, resources.length);
-    total += resources.length;
-    nextToken = response.NextToken;
-  } while (nextToken);
-  AWS_MAPPING.total += total;
-};
 
-const queryLambdaFunction = async (serviceName, resourceType, region) => {
-  let resources = [];
-  const client = new LambdaClient({ region });
-  for await (const page of paginateListFunctions({ client }, {})) {
-    resources.push(...(page.Functions || []));
-  }
-  const resourceCount = resources.length;
-  updateResourceTypeCounter(serviceName, resourceType, resourceCount);
-  AWS_MAPPING.total += resourceCount;
-};
 
-const queryKinesisStream = async (serviceName, resourceType, region) => {
-  let total = 0;
-  const client = new KinesisClient({ region });
-  const resources = [];
-  let hasMoreStreams = false;
-  do {
-    const command = new ListStreamsCommand({
-      ExclusiveStartStreamName:
-        resources.length > 0 ? resources[resources.length - 1] : undefined
-    });
-    const response = await client.send(command);
 
-    resources.push(...(response.StreamNames || []));
-    updateResourceTypeCounter(serviceName, resourceType, resources.length);
-    total += resources.length;
-    hasMoreStreams = response.HasMoreStreams;
-  } while (hasMoreStreams);
-  AWS_MAPPING.total += total;
-};
 
-const queryKinesisFirehoseDeliveryStream = async (
-  serviceName,
-  resourceType,
-  region
-) => {
-  let total = 0;
-  const client = new FirehoseClient({ region });
-  const resources = [];
-  let hasMoreDeliveryStreams = false;
-  do {
-    const command = new ListDeliveryStreamsCommand({
-      ExclusiveStartDeliveryStreamName:
-        resources.length > 0 ? resources[resources.length - 1] : undefined
-    });
-    const response = await client.send(command);
 
-    resources.push(...(response.DeliveryStreamNames || []));
-    updateResourceTypeCounter(serviceName, resourceType, resources.length);
-    total += resources.length;
-    hasMoreDeliveryStreams = response.HasMoreDeliveryStreams;
-  } while (hasMoreDeliveryStreams);
-  AWS_MAPPING.total += total;
-};
 
-const queryIAMAccountAuthorization = async (
-  serviceName,
-  resourceType,
-  region
-) => {
-  const users = [];
-  const groups = [];
-  const roles = [];
-  const managedPolicies = [];
-  if (
-    AWS_MAPPING[serviceName] &&
-    AWS_MAPPING[serviceName][resourceType] !== undefined
-  ) {
-    return;
-  }
-  const client = new IAMClient({ region });
-  for await (const page of paginateGetAccountAuthorizationDetails(
-    { client },
-    {}
-  )) {
-    if (page.UserDetailList && page.UserDetailList.length) {
-      users.push(...(page.UserDetailList || []));
-      updateResourceTypeCounter(
-        serviceName,
-        IAM_USER,
-        page.UserDetailList.length
-      );
-    } else {
-      updateResourceTypeCounter(serviceName, IAM_USER, 0);
-    }
-    if (page.GroupDetailList && page.GroupDetailList.length) {
-      groups.push(...(page.GroupDetailList || []));
-      updateResourceTypeCounter(
-        serviceName,
-        IAM_GROUP,
-        page.GroupDetailList.length
-      );
-    } else {
-      updateResourceTypeCounter(serviceName, IAM_GROUP, 0);
-    }
-    if (page.RoleDetailList && page.RoleDetailList.length) {
-      roles.push(...(page.RoleDetailList || []));
-      updateResourceTypeCounter(
-        serviceName,
-        IAM_ROLE,
-        page.RoleDetailList.length
-      );
-    } else {
-      updateResourceTypeCounter(serviceName, IAM_ROLE, 0);
-    }
-    if (page.Policies && page.Policies.length) {
-      managedPolicies.push(...(page.Policies || []));
-      updateResourceTypeCounter(serviceName, IAM_POLICY, page.Policies.length);
-    } else {
-      updateResourceTypeCounter(serviceName, IAM_POLICY, 0);
-    }
-  }
-  const total =
-    users.length + groups.length + roles.length + managedPolicies.length;
-  AWS_MAPPING.total += total;
-};
 
-const queryGlacierVault = async (serviceName, resourceType, region) => {
-  let total = 0;
-  let resources = [];
-  const client = new GlacierClient({ region });
-  for await (const page of paginateListVaults({ client }, {})) {
-    resources.push(...(page.VaultList || []));
-  }
-  const resourceCount = resources.length;
-  total += resourceCount;
-  updateResourceTypeCounter(serviceName, resourceType, resourceCount);
-  AWS_MAPPING.total += total;
-};
 
-const queryWafv2RuleGroup = async (serviceName, resourceType, region) => {
-  let total = 0;
-  const client = new WAFV2Client({ region });
-  const responseForRegional = await paginate({
-    client,
-    command: {
-      CommandClass: ListRuleGroupsCommand,
-      params: {
-        Scope: "REGIONAL"
-      }
-    },
-    nextKey: "NextMarker",
-    responseKey: "RuleGroups"
-  });
 
-  // CLOUDFRONT only supported for us-east-1
-  let responseForCloudfront;
-  if (region === "us-east-1") {
-    responseForCloudfront = await paginate({
-      client,
-      command: {
-        CommandClass: ListRuleGroupsCommand,
-        params: {
-          Scope: "CLOUDFRONT"
-        }
-      },
-      nextKey: "NextMarker",
-      responseKey: "RuleGroups"
-    });
-  }
 
-  const data = [];
-  if (responseForRegional.RuleGroups)
-    data.push(...responseForRegional.RuleGroups);
 
-  if (responseForCloudfront.RuleGroups)
-    data.push(...responseForCloudfront.RuleGroups);
-  const resourceCount = data.length;
-  total += resourceCount;
-  updateResourceTypeCounter(serviceName, resourceType, resourceCount);
-  AWS_MAPPING.total += total;
-};
 
-const queryEc2Ami = async (serviceName, resourceType, region) => {
-  const client = new EC2Client({ region });
-  let resources = [];
-  for await (const image of paginateDescribeImages(
-    { client },
-    { Owners: ["self"] }
-  )) {
-    const { Images: images } = image;
-    resources.push(...(images || []));
-  }
-  updateResourceTypeCounter(serviceName, resourceType, resources.length);
-  AWS_MAPPING.total += resources.length;
-};
 
-const queryEc2fpga = async (serviceName, resourceType, region) => {
-  const client = new EC2Client({ region });
-  let resources = [];
-  for await (const image of paginateDescribeFpgaImages(
-    { client },
-    { Owners: ["self"] }
-  )) {
-    const { FpgaImages: images } = image;
-    resources.push(...(images || []));
-  }
-  updateResourceTypeCounter(serviceName, resourceType, resources.length);
-  AWS_MAPPING.total += resources.length;
-};
 
-const queryScalableTargets = async (serviceName, resourceType, region) => {
-  const aasClient = new ApplicationAutoScalingClient({ region });
-  let total = 0;
-  let aasNextToken;
-  do {
-    const describeScalableTargets = new DescribeScalableTargetsCommand({
-      NextToken: aasNextToken,
-      ServiceNamespace: "dynamodb"
-    });
-    const response = await aasClient.send(describeScalableTargets);
-    if (response && response.ScalableTargets) {
-      const scalableTargetsCount = response.ScalableTargets.length || 0;
-      total += scalableTargetsCount;
-      updateResourceTypeCounter(
-        serviceName,
-        resourceType,
-        scalableTargetsCount
-      );
-      aasNextToken = response.NextToken;
-    }
-  } while (aasNextToken);
-  AWS_MAPPING.total += total;
-};
 
-const queryEC2Instance = async (serviceName, resourceType, region) => {
-  let total = 0;
-  let resources = [];
-  try {
-    const client = new EC2Client({ region });
-    for await (const page of paginateDescribeInstances({ client }, {})) {
-      if (page.Reservations) {
-        resources.push(
-          ...(page.Reservations.map(
-            (reservation) => reservation.Instances
-          ).flat() || [])
-        );
-      }
-    }
-    const ec2InstancesCount = resources.length;
-    total += ec2InstancesCount;
-    updateResourceTypeCounter(serviceName, resourceType, ec2InstancesCount);
-  } catch (err) {
-    console.log(`Error finding ${resourceType}`);
-  }
-  AWS_MAPPING.total += total;
-};
 
-const querySnapshot = async (serviceName, resourceType, region) => {
-  const client = new EC2Client({ region });
-  let resources = [];
-  for await (const image of paginateDescribeSnapshots(
-    { client },
-    { OwnerIds: ["self"] }
-  )) {
-    const { Snapshots:  snapshots } = snapshots;
-    resources.push(...(snapshots || []));
-  }
-  updateResourceTypeCounter(serviceName, resourceType, resources.length);
-  AWS_MAPPING.total += resources.length;
-};
 
-const queryDefaultEBSEncryption = async (serviceName, resourceType, region) => {
-  let total = 0;
-  const command = new GetEbsEncryptionByDefaultCommand({});
-  const client = new EC2Client({ region });
-  const response = await client.send(command);
-  if (response) {
-    const ebsEncryptionCount =
-      response.EbsEncryptionByDefault !== undefined ? 1 : 0;
-    total += ebsEncryptionCount;
-    updateResourceTypeCounter(serviceName, resourceType, ebsEncryptionCount);
-  }
 
-  AWS_MAPPING.total += total;
-};
 
-const queryAPIGatewayStage = async (serviceName, resourceType, region) => {
-  let total = 0;
-  const client = new APIGatewayClient({ region });
-  const resources = [];
-  for await (const page of paginateGetRestApis({ client }, {})) {
-    resources.push(...(page.items || []));
-  }
-  for await (const apis of resources) {
-    const command = new GetStagesCommand({
-      restApiId: apis.id
-    });
-    const response = await client?.send(command);
-    const stageCount = response?.item.length;
-    total += stageCount;
-    updateResourceTypeCounter(serviceName, resourceType, stageCount);
-  }
-  AWS_MAPPING.total += total;
-};
 
-const queryECSService = async (serviceName, resourceType, region) => {
-  let total = 0;
-  const client = new ECSClient({ region });
-  const resources = [];
-  for await (const page of paginateListClusters({ client }, {})) {
-    resources.push(...(page.clusterArns || []));
-  }
 
-  for await (const clusterArn of resources) {
-    const command = new ListServicesCommand({
-      cluster: clusterArn
-    });
-    const response = await client?.send(command);
-    const serviceCount = response?.serviceArns.length;
-    total += serviceCount;
-    updateResourceTypeCounter(serviceName, resourceType, serviceCount);
-    const chunkedServices = chunks(response?.serviceArns, 10);
-    // Batching by 10 because this is the API limit per docs
-    // https://docs.aws.amazon.com/AWSJavaScriptSDK/v3/latest/clients/client-ecs/interfaces/describeservicescommandinput.html
-    await Promise.all(
-      chunkedServices.map(async (serviceArns) => {
-        const command = new DescribeServicesCommand({
-          cluster: clusterArn,
-          services: serviceArns
-        });
-        const response = await client?.send(command);
-        const taskDefinitionArns = [];
-        if (response && Array.isArray(response?.services)) {
-          response.services.forEach((service) => {
-            if (service?.taskDefinition) {
-              taskDefinitionArns.push(service.taskDefinition);
-            }
-            if (Array.isArray(service?.deployments)) {
-              const deploymentTaskDefinitionArns = service.deployments
-                .filter(
-                  (deployment) =>
-                    deployment?.status === "PRIMARY" &&
-                    !!deployment?.taskDefinition
-                )
-                .map((deployment) => deployment?.taskDefinition);
-              if (deploymentTaskDefinitionArns) {
-                taskDefinitionArns.push(...deploymentTaskDefinitionArns);
-              }
-            }
-          });
-        }
 
-        await Promise.all(
-          [...new Set(taskDefinitionArns)].map(async (taskDefinition) => {
-            let getTaskDefCmd = new DescribeTaskDefinitionCommand({
-              taskDefinition
-            });
-            let getTaskDefRes = await client?.send(getTaskDefCmd);
-            let containerCount =
-              getTaskDefRes.taskDefinition?.containerDefinitions.length;
-            total += containerCount;
-            updateResourceTypeCounter(
-              serviceName,
-              ECS_TASK_DEFINITION,
-              containerCount
-            );
-          })
-        );
-      })
-    );
-  }
-  AWS_MAPPING.total += total;
-};
 
-const queryDependencies = async (serviceName, resourceType, region) => {
-  const resourceDependency = dependencies.find((dep) => dep.id === serviceName);
-  let total = 0;
-  if (resourceDependency) {
-    const { resources } = resourceDependency;
-    const resourceTypeDependency = resources.find(
-      (resource) => resource.resourceType === resourceType
-    );
-    if (resourceTypeDependency) {
-      if (
-        resourceTypeDependency.listClientConfig.clientType ===
-        "AWS_CLOUDCONTROL_V3"
-      ) {
-        const listClient = new CloudControlClient({
-          region
-        });
-        const paginatorConfig = {
-          client: listClient,
-          pageSize: 25
-        };
-        const commandInput = {
-          TypeName: resourceType
-        };
-        const paginator = paginateListResources(paginatorConfig, commandInput);
-        for await (const page of paginator) {
-          const descriptors = page.ResourceDescriptions;
-          const s3BucketsLength = descriptors.length || 0;
-          updateResourceTypeCounter(serviceName, resourceType, s3BucketsLength);
-          total += s3BucketsLength;
-        }
-      } else {
-        const { list, listClientConfig } = resourceTypeDependency;
-        const classPackage = await import(listClientConfig.packageName);
-        const ClientClass = classPackage[listClientConfig.clientName];
-        const listClient = new ClientClass({
-          region
-        });
-        const { action: listAction } = list;
-        const { command, outputProperty, outputPropertySearchType } =
-          listAction;
 
-        const resources = [];
-        if (command.includes("paginate")) {
-          const packageImport = await import(listClientConfig.packageName);
-          const CommandClass = packageImport[command];
-          for await (const page of CommandClass({ client: listClient }, {})) {
-            if (outputPropertySearchType === "JSON_PATH") {
-              const result = jp.query(page, outputProperty)[0];
-              resources.push(...result);
-            } else {
-              resources.push(...page[outputProperty]);
-            }
-          }
-        } else {
-          const packageImport = await import(listClientConfig.packageName);
-          const CommandClass = packageImport[`${command}Command`];
-          const com = new CommandClass({});
-          let nextResponsePageToken;
-          let nextRequestPageToken;
-          let nextPageTokenValue;
-          do {
-            const paginationArgs =
-              nextResponsePageToken && nextPageTokenValue && nextRequestPageToken
-                ? { [nextRequestPageToken]: nextPageTokenValue }
-                : {};
-            const com = new CommandClass(paginationArgs);
-            // @ts-ignore
-            const listResponse = await listClient?.send(com);
-            nextResponsePageToken = getNextPageTokenKeyFromResponse(
-              listResponse,
-              getNextListPageTokenKey(list)
-            );
-            nextPageTokenValue =
-              nextResponsePageToken && listResponse[nextResponsePageToken];
-            if (outputPropertySearchType === "JSON_PATH") {
-              const result = jp.query(listResponse, outputProperty)[0];
-              const responses = Array.isArray(result) ? result : [result];
-              resources.push(...responses);
-            } else {
-              const responses = Array.isArray(listResponse[outputProperty])
-                ? listResponse[outputProperty]
-                : [listResponse[outputProperty]];
-              resources.push(...responses);
-            }
-            nextRequestPageToken = getNextListPageTokenRequestKey(
-              nextResponsePageToken,
-              list
-            );
-          } while (
-            nextPageTokenValue &&
-            nextResponsePageToken &&
-            nextRequestPageToken
-            );
-        }
-        const resultsLength = resources.length || 0;
-        updateResourceTypeCounter(serviceName, resourceType, resultsLength);
-        total += resultsLength;
-      }
-    }
-  }
-  AWS_MAPPING.total += total;
-};
 
-const updateResourceTypeCounter = (serviceName, resourceType, value) => {
-  if (AWS_MAPPING[serviceName] === undefined) {
-    AWS_MAPPING[serviceName] = { [resourceType]: value };
-  } else if (AWS_MAPPING[serviceName][resourceType] === undefined) {
-    AWS_MAPPING[serviceName][resourceType] = value;
-  } else {
-    AWS_MAPPING[serviceName][resourceType] += value;
-  }
-};
 
-const queryDocDBInstance = async (serviceName, resourceType, region) => {
-  let resources = [];
-  const client = new DocDBClient({ region });
-  for await (const page of paginateDescribeDBInstances(
-    { client },
-    { Filters: [{ Name: "engine", Values: ["docdb"] }] }
-  )) {
-    resources.push(...(page.DBInstances || []));
-  }
-  const resourceCount = resources.length;
-  updateResourceTypeCounter(serviceName, resourceType, resourceCount);
-  AWS_MAPPING.total += resourceCount;
-};
 
-const queryDocDBCluster = async (serviceName, resourceType, region) => {
-  let resources = [];
-  const client = new DocDBClient({ region });
-  for await (const page of paginateDescribeDBClusters(
-    { client },
-    { Filters: [{ Name: "engine", Values: ["docdb"] }] }
-  )) {
-    resources.push(...(page.DBClusters || []));
-  }
-  const resourceCount = resources.length;
-  updateResourceTypeCounter(serviceName, resourceType, resourceCount);
-  AWS_MAPPING.total += resourceCount;
-};
 
-const queryDocDBClusterParameterGroup = async (serviceName, resourceType, region) => {
-  let resources = [];
-  const client = new DocDBClient({ region });
-  for await (const page of paginateDocDBClusterParameterGroups(
-    { client },
-    { Filters: [{ Name: "engine", Values: ["docdb"] }] }
-  )) {
-    resources.push(...(page.DBClusterParameterGroups || []));
-  }
-  const resourceCount = resources.length;
-  updateResourceTypeCounter(serviceName, resourceType, resourceCount);
-  AWS_MAPPING.total += resourceCount;
-};
 
-const queryRDSInstance = async (serviceName, resourceType, region) => {
-  let resources = [];
-  const client = new RDSClient({ region });
-  for await (const page of paginateRDSInstances(
-    { client },
-    {
-      Filters: [
-        {
-          Name: "engine",
-          Values: [
-            "aurora-mysql",
-            "custom-sqlserver-ee",
-            "custom-sqlserver-se",
-            "custom-sqlserver-web",
-            "aurora-postgresql",
-            "mariadb",
-            "mysql",
-            "oracle-ee",
-            "oracle-ee-cdb",
-            "oracle-se2",
-            "oracle-se2-cdb",
-            "aurora",
-            "postgres",
-            "sqlserver-ee",
-            "sqlserver-ex",
-            "sqlserver-se",
-            "sqlserver-web"
-          ]
-        }
-      ]
-    }
-  )) {
-    resources.push(...(page.DBInstances || []));
-  }
-  const resourceCount = resources.length;
-  updateResourceTypeCounter(serviceName, resourceType, resourceCount);
-  AWS_MAPPING.total += resourceCount;
-};
 
-const queryRDSCluster = async (serviceName, resourceType, region) => {
-  let resources = [];
-  const client = new RDSClient({ region });
-  for await (const page of paginateRDSClusters(
-    { client },
-    {
-      Filters: [
-        {
-          Name: "engine",
-          Values: [
-            "aurora-mysql",
-            "custom-sqlserver-ee",
-            "custom-sqlserver-se",
-            "custom-sqlserver-web",
-            "aurora-postgresql",
-            "mariadb",
-            "mysql",
-            "oracle-ee",
-            "oracle-ee-cdb",
-            "oracle-se2",
-            "oracle-se2-cdb",
-            "aurora",
-            "postgres",
-            "sqlserver-ee",
-            "sqlserver-ex",
-            "sqlserver-se",
-            "sqlserver-web"
-          ]
-        }
-      ]
-    }
-  )) {
-    resources.push(...(page.DBClusters || []));
-  }
-  const resourceCount = resources.length;
-  updateResourceTypeCounter(serviceName, resourceType, resourceCount);
-  AWS_MAPPING.total += resourceCount;
-};
 
-const queryRDSClusterSnapshot = async (serviceName, resourceType, region) => {
-  let resources = [];
-  const client = new RDSClient({ region });
-  for await (const page of paginateDescribeDBClusterSnapshots(
-    { client },
-    {
-      Filters: [
-        {
-          Name: "engine",
-          Values: [
-            "aurora-mysql",
-            "custom-sqlserver-ee",
-            "custom-sqlserver-se",
-            "custom-sqlserver-web",
-            "aurora-postgresql",
-            "mariadb",
-            "mysql",
-            "oracle-ee",
-            "oracle-ee-cdb",
-            "oracle-se2",
-            "oracle-se2-cdb",
-            "aurora",
-            "postgres",
-            "sqlserver-ee",
-            "sqlserver-ex",
-            "sqlserver-se",
-            "sqlserver-web"
-          ]
-        }
-      ]
-    }
-  )) {
-    resources.push(...(page.DBClusterSnapshots || []));
-  }
-  const resourceCount = resources.length;
-  updateResourceTypeCounter(serviceName, resourceType, resourceCount);
-  AWS_MAPPING.total += resourceCount;
-};
 
-const queryEKSNodeGroup = async (serviceName, resourceType, region) => {
-  const client = new EKSClient({ region });
-  const clusterNames = [];
-  for await (const page of paginateListEKSClusters({ client }, {})) {
-    clusterNames.push(...(page.clusters || []));
-  }
-  let nodeGroups = [];
-  await Promise.all(
-    clusterNames.map(async (clusterName) => {
-      for await (const page of paginateListNodegroups(
-        { client },
-        { clusterName }
-      )) {
-        if (
-          page?.nodegroups?.length !== undefined &&
-          page?.nodegroups?.length > 0
-        ) {
-          for (const nodeGroup of page.nodegroups) {
-            nodeGroups.push({ clusterName, nodeGroup });
-          }
-        }
-      }
-    })
-  );
-  const resourceCount = nodeGroups.length;
-  updateResourceTypeCounter(serviceName, resourceType, resourceCount);
-  AWS_MAPPING.total += resourceCount;
-};
 
-const queryNeptuneInstance = async (serviceName, resourceType, region) => {
-  let resources = [];
-  const client = new NeptuneClient({ region });
-  for await (const page of paginateNeptuneInstances(
-    { client },
-    { Filters: [{ Name: "engine", Values: ["neptune"] }] }
-  )) {
-    resources.push(...(page.DBInstances || []));
-  }
-  const resourceCount = resources.length;
-  updateResourceTypeCounter(serviceName, resourceType, resourceCount);
-  AWS_MAPPING.total += resourceCount;
-};
 
-const queryNeptuneCluster = async (serviceName, resourceType, region) => {
-  let resources = [];
-  const client = new NeptuneClient({ region });
-  for await (const page of paginateNeptuneClusters(
-    { client },
-    { Filters: [{ Name: "engine", Values: ["neptune"] }] }
-  )) {
-    resources.push(...(page.DBClusters || []));
-  }
-  const resourceCount = resources.length;
-  updateResourceTypeCounter(serviceName, resourceType, resourceCount);
-  AWS_MAPPING.total += resourceCount;
-};
-
-const queryNeptuneClusterParameterGroups = async (
-  serviceName,
-  resourceType,
-  region
-) => {
-  let resources = [];
-  const client = new NeptuneClient({ region });
-  for await (const page of paginateDescribeDBClusterParameterGroups(
-    { client },
-    { Filters: [{ Name: "engine", Values: ["neptune"] }] }
-  )) {
-    resources.push(...(page.DBClusterParameterGroups || []));
-  }
-  const resourceCount = resources.length;
-  updateResourceTypeCounter(serviceName, resourceType, resourceCount);
-  AWS_MAPPING.total += resourceCount;
-};
-
-const queryComputeOptimizerEnrollmentStatus = async (
-  serviceName,
-  resourceType,
-  region
-) => {
-  let total = 0;
-  const client = new ComputeOptimizerClient({ region });
-  const resources = [];
-  const command = new GetEnrollmentStatusCommand({});
-  const response = await client.send(command);
-  resources.push({ ...(response || {}) });
-  updateResourceTypeCounter(serviceName, resourceType, resources.length);
-  total += resources.length;
-  AWS_MAPPING.total += total;
-};
-
-const queryScalingPolicies = async (serviceName, resourceType, region) => {
-  let total = 0;
-  const resources = [];
-  const client = new ApplicationAutoScalingClient({ region });
-  for await (const page of paginateDescribeScalingPolicies(
-    { client },
-    { ServiceNamespace: "ecs" }
-  )) {
-    resources.push(...(page.ScalingPolicies || []));
-  }
-  const resourceCount = resources.length;
-  updateResourceTypeCounter(serviceName, resourceType, resourceCount);
-  total += resources.length;
-  AWS_MAPPING.total += total;
-};
-
-const queryS3Buckets = async (serviceName, resourceType, region) => {
-  let total = 0;
-  const resources = [];
-  const client = new S3Client({ region });
-  const command = new ListBucketsCommand({});
-  const response = await client.send(command);
-  resources.push({ ...(response.Buckets || []) });
-  updateResourceTypeCounter(serviceName, resourceType, resources.length);
-  total += resources.length;
-  AWS_MAPPING.total += total;
-};
-
-const queryEMRClusters = async (serviceName, resourceType, region) => {
-  let total = 0;
-  const resources = [];
-  const client = new EMRClient({ region });
-  for await (const page of paginateListClusters(
-    { client },
-    {}
-  )) {
-    resources.push(...(page.Clusters || []));
-  }
-  const resourceCount = resources.length;
-  updateResourceTypeCounter(serviceName, resourceType, resourceCount);
-  total += resources.length;
-  AWS_MAPPING.total += total;
-};
-
-const queryAPIGatewayV2APIs = async (serviceName, resourceType, region) => {
-  let total = 0;
-  const resources = [];
-  const client = new ApiGatewayV2Client({ region });
-  let nextToken;
-  do {
-    const command = new GetApisCommand({
-      NextToken: nextToken
-    });
-    const response = await client.send(command);
-    resources.push(...(response.Items || []));
-    nextToken = response.NextToken;
-  } while (nextToken);
-  const resourceCount = resources.length;
-  updateResourceTypeCounter(serviceName, resourceType, resourceCount);
-  total += resources.length;
-  AWS_MAPPING.total += total;
-};
 
 export const queryAWS = async (parsedService, parsedResourceType) => {
   await Promise.all(
